Move EmptyPageProvider into test namespace
